Refactor amodule and dispatcher filename
