+++ hydrabus_framework/core/command/run.py
@@ -1,6 +1,6 @@
 import traceback
 
-from hydrabus_framework.modules.amodule import AModule
+from hydrabus_framework.modules.AModule import AModule
 
 
 __author__ = "Jordan Ovrè <ghecko78@gmail.com>"
